Added a "delay_range" parameter to the click method
This delay_range parameter is a tuple that expects two values; one float for a lower bound and another float for an upper bound. A random float between the lower bound and upper bound will be used as a sleep duration before the mouse is clicked. This is to better simulate human behaviour. By default there is no delay range.